+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Upcoming changes ...
+### Added
+- Added read/write stream support to read scan raw results
 
 ## [1.27.0] - 2025-12-12
 ### Added