+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed pre selected license in inventory dialog
 - Fixed local cryptography scanning to include all project files, excluding only configuration files at the project root
+### Changed
+- Upgrades `scanoss.js` SDK version to v0.29.0
 
 ## [1.27.0] - 2025-12-12
 ### Added