+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Upcoming changes ...
+
+## [1.28.0] - 2025-12-22
 ### Added
 - Added read/write stream support to read scan raw results
 - Added virtualization in cryptography viewer component
@@ -81,4 +83,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.26.0]: https://github.com/scanoss/sbom-workbench/compare/v1.25.0...v1.26.0
 [1.26.1]: https://github.com/scanoss/sbom-workbench/compare/v1.26.0...v1.26.1
 [1.27.0]: https://github.com/scanoss/sbom-workbench/compare/v1.26.1...v1.27.0
+[1.28.0]: https://github.com/scanoss/sbom-workbench/compare/v1.27.0...v1.28.0